# specs/ticket-service.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ticket Service API",
    "description": "Спецификация сервиса управления билетами и сервиса бронирования.\nВсе параметры операций передаются в URL. В POST, PUT и PATCH данные Ticket\nпередаются в query-параметре `ticket` как JSON.\nПоля id, creationDate и event.id генерируются сервером.\n",
    "version": "1.0.0",
    "contact": {
      "name": "Nostya && Dasha"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Основной сервис управления коллекцией билетов"
    }
  ],
...
  "paths": {
...
    "/tickets/max-type": {
      "get": {
        "tags": [
          "Ticket queries"
        ],
        "summary": "Получить билет с максимальным типом",
        "description": "Возвращает один любой билет, поле `type` которого имеет\nмаксимальное значение среди всех билетов коллекции.\n\nЗначения сравниваются в порядке объявления перечисления:\n`VIP < USUAL < BUDGETARY < CHEAP`.\n\nЕсли максимальный тип имеют несколько билетов, сервис может\nвернуть любой из них.\n",
        "operationId": "getTicketWithMaxType",
        "responses": {
          "200": {
            "description": "Билет с максимальным значением поля type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "404": {
            "description": "Коллекция билетов пуста, поэтому определить билет с максимальным type невозможно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Коллекция билетов пуста.",
                  "path": "/api/v1/tickets/max-type"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/comment-contains": {
      "get": {
        "tags": [
          "Ticket queries"
        ],
        "summary": "Найти билеты по подстроке в комментарии",
        "description": "Возвращает массив билетов, поле \"comment\" которых содержит\nзаданную подстроку\n",
        "operationId": "getTicketsByCommentSubstring",
        "parameters": [
          {
            "name": "substring",
            "in": "query",
            "required": true,
            "description": "Подстрока, которую должно содержать поле comment",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "example": "сцена"
          }
        ],
        "responses": {
          "200": {
            "description": "Массив найденных билетов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Обязательный параметр substring отсутствует или является пустой строкой.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "substring должен содержать хотя бы один символ.",
                  "path": "/api/v1/tickets/comment-contains",
                  "violations": [
                    {
                      "field": "substring",
                      "message": "substring должен содержать хотя бы один символ."
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/tickets/type-greater-than": {
      "get": {
        "tags": [
          "Ticket queries"
        ],
        "summary": "Получить билеты с типом больше заданного",
        "description": "Возвращает массив билетов, значение поля `type` которых больше\nпереданного значения\n",
        "operationId": "getTicketsWithTypeGreaterThan",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "description": "Тип билета, с которым выполняется сравнение",
            "schema": {
              "$ref": "#/components/schemas/TicketType"
            },
            "example": "USUAL"
          }
        ],
        "responses": {
          "200": {
            "description": "Массив билетов с типом больше заданного",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Обязательный параметр type отсутствует или не равен VIP, USUAL, BUDGETARY либо CHEAP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "type должен быть VIP, USUAL, BUDGETARY или CHEAP; получено STANDARD.",
                  "path": "/api/v1/tickets/type-greater-than",
                  "violations": [
                    {
                      "field": "type",
                      "message": "type должен быть VIP, USUAL, BUDGETARY или CHEAP; получено STANDARD."
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ticket": {
        "type": "object",
        "description": "Билет",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "readOnly": true,
            "description": "Уникальный автоматически генерируемый идентификатор билета",
            "example": 10
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Название билета. Строка не может быть пустой",
            "example": "Билет в партер"
          },
          "coordinates": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "creationDate": {
            "type": "string",
            "pattern": "^(0[1-9]|[12][0-9]|3[01])\\.(0[1-9]|1[0-2])\\.\\d{4} ([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\\.\\d{1,9})?$",
            "readOnly": true,
            "description": "Дата и время создания, генерируются сервером. Формат dd.MM.yyyy HH:mm:ss с необязательными долями секунды (1–9 цифр), без часового пояса. Пример: 21 сентября 2026, 14:30. Календарную корректность проверяет сервер.",
            "example": "21.09.2026 14:30:00"
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Цена билета. Значение должно быть больше 0",
            "example": 2500.5
          },
          "comment": {
            "type": "string",
            "maxLength": 341,
            "description": "Комментарий к билету. Поле не может быть null",
            "example": "Место рядом со сценой"
          },
          "type": {
            "$ref": "#/components/schemas/TicketType"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        },
        "required": [
          "id",
          "name",
          "coordinates",
          "creationDate",
          "price",
          "comment",
          "type"
        ],
        "title": "Билет"
      },
...
      "Coordinates": {
        "type": "object",
        "description": "Координаты билета",
        "properties": {
          "x": {
            "type": "number",
            "format": "double",
            "minimum": -239,
            "exclusiveMinimum": true,
            "description": "Координата X (> -239)",
            "example": 10.5
          },
          "y": {
            "type": "integer",
            "format": "int64",
            "minimum": -495,
            "exclusiveMinimum": true,
            "description": "Координата Y (> -495)",
            "example": 20
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "Координаты"
      },
      "Event": {
        "type": "object",
        "nullable": true,
        "description": "Мероприятие, для которого предназначен билет",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "readOnly": true,
            "description": "Уникальный автоматически генерируемый идентификатор мероприятия, должен быть >0",
            "example": 7
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Название мероприятия. Строка не может быть пустой",
            "example": "Вечер оперы"
          },
          "ticketsCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "description": "Количество билетов. Значение должно быть больше 0",
            "example": 120
          },
          "eventType": {
            "$ref": "#/components/schemas/EventType"
          }
        },
        "required": [
          "id",
          "name",
          "ticketsCount"
        ],
        "title": "Мероприятие"
      },
      "TicketType": {
        "type": "string",
        "description": "Тип билета. Порядок сравнения: VIP < USUAL < BUDGETARY < CHEAP",
        "enum": [
          "VIP",
          "USUAL",
          "BUDGETARY",
          "CHEAP"
        ],
        "title": "Тип билета"
      },
      "EventType": {
        "type": "string",
        "nullable": true,
        "description": "Тип мероприятия",
        "enum": [
          "CONCERT",
          "BASKETBALL",
          "OPERA",
          null
        ],
        "title": "Тип мероприятия"
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Информация об ошибке выполнения запроса",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Время ошибки в UTC в формате dd.MM.yyyy HH:mm:ss, необязательные доли секунды.",
            "pattern": "^(0[1-9]|[12][0-9]|3[01])\\.(0[1-9]|1[0-2])\\.\\d{4} ([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\\.\\d{1,9})?$",
            "example": "21.09.2026 14:30:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP-код ошибки",
            "minimum": 400,
            "maximum": 599,
            "example": 422
          },
          "error": {
            "type": "string",
            "description": "Краткое название ошибки",
            "example": "Unprocessable Content"
          },
          "message": {
            "type": "string",
            "description": "Подробное описание причины ошибки",
            "example": "price должен быть больше 0; получено -10."
          },
          "path": {
            "type": "string",
            "description": "URL запроса, при обработке которого произошла ошибка",
            "example": "/api/v1/tickets"
          },
          "violations": {
            "type": "array",
            "description": "Список нарушений схемы с именами полей. Для ошибок разбора может указываться весь параметр ticket.",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Путь поля или имя параметра.",
                  "example": "ticket.price"
                },
                "message": {
                  "type": "string",
                  "description": "Условие, которое нарушено, и переданное значение.",
                  "example": "price должен быть больше 0; получено -10."
                }
              }
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "message",
          "error",
          "path"
        ],
        "title": "Ошибка API"
      },
...
    },
...
  }
}